projects/requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
bitarray==1.0.1
bleach==3.1.4
blinker==1.4
boltons==19.1.0
Brotli==1.0.7
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
crawlab==0.1.0a8
crawlab-sdk==0.1.6
crochet==1.12.0
cryptography==2.4.2
cssselect==1.1.0
decorator==4.4.2
defusedxml==0.6.0
demjson==2.2.4
Django==2.2.4
django-cors-headers==3.1.0
docker==4.1.0
docker-py==1.10.6
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.3
Fabric3==1.14.post1
face==19.1.2
Flask==1.1.1
fonttools==4.0.2
gerapy==0.8.5
gevent==1.4.0
glom==19.2.0
greenlet==0.4.15
h11==0.9.0
h2==3.2.0
hpack==3.0.0
html5lib==1.0.1
HTMLParser==0.0.2
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.2
incremental==17.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.1
jiphy==1.2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kaitaistruct==0.8
ldap3==2.6.1
lxml==4.4.1
MarkupSafe==1.1.1
mistune==0.8.4
mitmproxy==5.0.1
mysqlclient==1.4.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.17.1
pandas==0.25.1
pandocfilters==1.4.2
paramiko==2.7.1
parsel==1.5.2
parso==0.6.2
passlib==1.7.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.1.0
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
Protego==0.1.16
protobuf==3.10.0
psutil==5.6.4
ptyprocess==0.6.0
publicsuffix2==2.20191221
pyasn1==0.4.6
pyasn1-modules==0.2.6
pybloom-live==3.0.0
pycparser==2.19
PyDispatcher==2.0.5
PyExecJS==1.5.1
Pygments==2.6.1
PyHamcrest==1.9.0
pymongo==3.10.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyperclip==1.7.0
pyrsistent==0.16.0
pytesseract==0.3.0
python-dateutil==2.8.0
python-owasp-zap-v2.4==0.0.14
python-scrapyd-api==2.1.2
pytidylib==0.3.2
pytz==2019.2
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
queuelib==1.5.0
redis==3.3.8
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Scrapy==1.8.0
scrapy-redis==0.6.8
scrapy-redis-bloomfilter==0.7.0
scrapy-splash==0.7.2
scrapydo==0.2.2
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.4
SQLAlchemy==1.3.8
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
treelib==1.5.5
Twisted==19.7.0
urllib3==1.25.3
urwid==2.0.1
validators==0.14.2
w3lib==1.21.0
wcwidth==0.1.9
web.py==0.40.dev0
webencodings==0.5.1
webpage==0.1
websocket-client==0.56.0
Werkzeug==0.16.0
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==0.14.1
xlrd==1.2.0
zipp==3.1.0
zope.interface==4.6.0
zstandard==0.12.0
